+++ nanodjango/__init__.py
@@ -1,4 +1,10 @@
-from .app import Django  # noqa
+import pluggy
+
+from .app import Django
+
+# Register pluggy hook
+hookimpl = pluggy.HookimplMarker("nanodjango")
 
 
 __version__ = "0.10.0"
+__all__ = ["Django", "hookimpl"]